init custom generator

+++ db/migrate/20230531204440_create_order_items.rb
@@ -2,7 +2,7 @@ class CreateOrderItems < ActiveRecord::Migration[7.0]
   def change
     create_table :order_items do |t|
       t.float :quantity
-      t.float :subtotal
+      t.decimal :subtotal
       t.references :product, null: false, foreign_key: true
       t.references :order, null: false, foreign_key: true
